Start with package restructure; Weapon now instantiated by class string name; Updated entity & player class to use getter and setters; Added mouseHandler button events, can now listen on Mouse button1 & button2;

+++ src/blockerino/util/AABB.java
@@ -13,12 +13,12 @@ public class AABB {
     private float radius;
     private int size;
 
-    public AABB(Vector2f _position, int _width, int _height) {
+    public AABB(Vector2f _position, float _width, float _height) {
         this.position = _position;
         this.width = _width;
         this.height = _height;
 
-        size = Math.max(_width, _height);
+        size = Math.max((int) _width, (int) _height);
     }
 
     public AABB(Vector2f _position, int _radius, Entity _entity) {
